feat: replace Double and Long in amount and price for BigDecimal

+++ src/main/kotlin/bass/dto/OrderDTO.kt
@@ -1,10 +1,11 @@
 package bass.dto
 
 import bass.entities.OrderEntity
+import java.math.BigDecimal
 
 class OrderDTO(
     val status: OrderEntity.OrderStatus,
-    val totalAmount: Double,
+    val totalAmount: BigDecimal,
     val memberId: Long,
     val items: List<OrderItemDTO> = mutableListOf(),
     val payments: List<PaymentDTO> = mutableListOf(),
